requirements.txt
python-osc>=1.8.0
pyautogui>=0.9.54
pyobjc-framework-Quartz>=9.0
pynput>=1.8.0
